Second commit
This file contains treatment of categorical variables, feature engineering and modelling. The categorical variables present in the data can be treated by One Hot Encoding or popularly known as dummy variable encoding. However, there are certain variables which appear to be numerical but are in fact categorical. Failing to treat such variables will lead to very poor accuracy and performance of the model. In this case it is the 'Hour' variable. 24hrs is divided into four categories with 6hrs in each and one hot encoding is applied to form 4 additional variables. The Hour column is then dropped off. In my project I wasn't able to understand why I was getting a poor accuracy of just 16% in Linear Regression model. After treating the 'Hour' variable however the model accuracy jumped to 80%.